+++ mobile/android/android-components/components/lib/dataprotect/src/main/java/mozilla/components/lib/dataprotect/KeySharedPreferences.kt
@@ -1,11 +1,10 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
+/* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 package mozilla.components.lib.dataprotect
 
+import android.annotation.TargetApi
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
 import android.os.Build
@@ -17,19 +16,47 @@ import java.security.GeneralSecurityException
 
 private const val BASE_64_FLAGS = Base64.URL_SAFE or Base64.NO_PADDING
 
-open class KeySharedPreferences(
+/**
+ * A wrapper around [SharedPreferences] which encrypts contents on supported API versions (23+).
+ * Otherwise, this simply delegates to [SharedPreferences].
+ *
+ * In rare circumstances (such as APK signing key rotation) a master key which protects this storage may be lost,
+ * in which case previously stored values will be lost as well. Applications are encouraged to instrument such events.
+ *
+ * @param context A [Context], used for accessing [SharedPreferences].
+ * @param keystore An optional [Keystore]: required when running on API23+, expected as `null` otherwise.
+ */
+class KeySharedPreferences(
     context: Context,
-    private val keystore: Keystore? = Keystore(BuildConfig.APPLICATION_ID)
+    private val keystore: Keystore? = null
 ) {
 
     companion object {
         const val KEY_PREFERENCES = "key_preferences"
     }
 
+    init {
+        if (Build.VERSION.SDK_INT < M) {
+            require(keystore == null) {
+                "Keystore isn't supported for pre-23 API versions"
+            }
+        } else {
+            require(keystore != null) {
+                "Keystore must be configured for 23+ API versions"
+            }
+        }
+    }
+
     private val logger = Logger("KeySharedPreferences")
     private val prefs = context.getSharedPreferences(KEY_PREFERENCES, MODE_PRIVATE)
 
-    open fun getString(key: String): String? {
+    /**
+     * Retrieves a stored [key]. See [putString] for storing a [key].
+     *
+     * @param key A key name.
+     * @return An optional [String] if [key] is present in the store.
+     */
+    fun getString(key: String): String? {
         return if (Build.VERSION.SDK_INT >= M) {
             getStringM(key)
         } else {
@@ -37,8 +64,33 @@ open class KeySharedPreferences(
         }
     }
 
+    /**
+     * Stores [value] under [key]. Retrieve it using [getString].
+     *
+     * @param key A key name.
+     * @param value A value for [key].
+     */
+    fun putString(key: String, value: String) {
+        if (Build.VERSION.SDK_INT >= M) {
+            putStringM(key, value)
+        } else {
+            prefs.edit().putString(key, value).apply()
+        }
+    }
+
+    /**
+     * Removes key/value pair from storage for the provided [key].
+     */
+    fun remove(key: String) {
+        prefs.edit().remove(key).apply()
+    }
+
+    @TargetApi(M)
     private fun getStringM(key: String): String? {
-        verifyKey()
+        // The fact that we're possibly generating a managed key here implies that this key could be lost after being
+        // for some reason. One possible reason for a key to be lost is rotating signing keys for the APK.
+        // Applications are encouraged to instrument such events.
+        generateManagedKeyIfNecessary()
 
         return if (prefs.contains(key)) {
             val value = prefs.getString(key, "")
@@ -62,16 +114,9 @@ open class KeySharedPreferences(
         }
     }
 
-    open fun putString(key: String, value: String) {
-        if (Build.VERSION.SDK_INT >= M) {
-            putStringM(key, value)
-        } else {
-            prefs.edit().putString(key, value).apply()
-        }
-    }
-
+    @TargetApi(M)
     private fun putStringM(key: String, value: String) {
-        verifyKey()
+        generateManagedKeyIfNecessary()
         val editor = prefs.edit()
 
         val encrypted = keystore!!.encryptBytes(value.toByteArray(StandardCharsets.UTF_8))
@@ -82,15 +127,11 @@ open class KeySharedPreferences(
         editor.putString(key, data).apply()
     }
 
-    open fun remove(key: String) {
-        val editor = prefs.edit()
-
-        editor.remove(key)
-
-        editor.apply()
-    }
-
-    private fun verifyKey() {
+    /**
+     * Generates a "managed key" - a key used to encrypt data stored by this class. This key is "managed" by [Keystore],
+     * which stores it in system's secure storage layer exposed via [AndroidKeyStore].
+     */
+    private fun generateManagedKeyIfNecessary() {
         if (!keystore!!.available()) {
             keystore.generateKey()
         }