+++ mobile/android/android-components/components/lib/dataprotect/src/test/java/mozilla/components/lib/dataprotect/KeystoreTest.kt
@@ -126,7 +126,7 @@ class KeystoreTest {
         Assert.assertNotNull(cipher)
     }
 
-    @Ignore("troubleshooting test-env crypto errors")
+    @Ignore("https://github.com/mozilla-mobile/android-components/issues/4956")
     @Test
     fun cryptoRoundTrip() {
         val keystore = Keystore("test-roundtrip", wrapper = wrapper)