+++ mobile/android/android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/Config.kt
@@ -8,6 +8,7 @@ import mozilla.components.concept.sync.DeviceCapability
 import mozilla.components.concept.sync.DeviceType
 import mozilla.components.service.fxa.manager.FxaAccountManager
 import mozilla.components.service.fxa.sync.GlobalSyncableStoreProvider
+import mozilla.components.lib.dataprotect.KeySharedPreferences
 
 typealias ServerConfig = mozilla.appservices.fxaclient.Config
 
@@ -66,7 +67,8 @@ sealed class SyncEngine(val nativeName: String) {
     object Bookmarks : SyncEngine("bookmarks")
 
     /**
-     * A 'logins/passwords' engine.
+     * A 'logins/passwords' engine. When using this engine, make sure to set an encryption key used to unlock the store
+     * as 'passwords_key' via [KeySharedPreferences].
      */
     object Passwords : SyncEngine("passwords")